Hanno's Magic
Stuff that was not allowing my computer to compile

+++ cine/simulation.cpp
@@ -380,7 +380,7 @@ namespace cine2 {
       }
     }
 
-	agents_.conflicts += conflicts_v.size();
+	agents_.conflicts += static_cast<int>(conflicts_v.size());
 
     conflicts_v.clear();
